vintrick-backend/requirements.txt
# vintrick-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
dotenv
pandas

